Refactor code. Address PR comments.
